Add a new class CancelableImpl that may be used by MamutFeatureComputer implementations to reduce code duplication
